refactor(data-java): use DateUtil to format Message, set PID to integer

+++ data-java/src/main/java/net/frakbot/crowdpulse/data/entity/ProjectRun.java
@@ -123,7 +123,7 @@ public class ProjectRun extends Entity {
      *
      * @param pid The {@link Double} PID.
      */
-    public void setPid(Double pid) {
+    public void setPid(Integer pid) {
         this.pid = pid;
     }
